[GLUON] Set proper location on restoring the insert point in gluon (#8531)
`warp_specialize` ops currently have unknown location set in the TTGIR
due to a quirk in the code emission in `_semantic.py`: for
`warp_specialize` we need save and then restore insert point. Location
is being inferred from the insert point, however if insert point happens
to be in a place that doesn't have location assigned (end of a block),
we set unknown loc. This change is a minimal fix that adds a helper that
gets the location from block's parent in such a case.
Alternatively we could also save location along with insert point, and
then restore it accordingly. This approach is simpler and should help
for most cases I could have think of however.
This change is important for consan changes I am working on, as it
breaks the LLVM backend if we create instrumentation function calls with
unknown location inferred from warp_specialize op.