[SWP] Dedup the code that checks if LoadOp can be converted to cpasync (#8529)
During SWP, we are checking if a given `LoadOp` should be lowered to
`AsyncCopyGlobalToLocalOp` twice - first in `AssignLatency`, and
`LowerLoops` next. The two checks duplicate non-trivial conditions like
`copyVecBytes >= 4` or `op.getResultTypes()[0].getIntOrFloatBitWidth()
>= 32`.

I moved the `isPipeliningBeneficial` function from `AssignLatency` into
utilities so that it can also be used by `LowerLoops`. This will also be
used by WS to determine if `LoadOp` should be lowered to cpasync and
assigned to the load partition.