[frontend] Disable cache when interpreter is enabled (#8499)
When `TRITON_INTERPRET=1`, the cache should always be disabled (in both
cases, `cache_results=true` and `TRITON_CACHE_AUTOTUNING=1`). Otherwise,
there is an error in `check_disk_cache` since it is searching for a
`JITFunction` which is not available.
This has been discussed in #6678 already but the introduced changes
still enable caching results when `cache_results=true`.