[FRONTEND] support multidimensional batches in tl.trans and tl.dot (#8542)
In particular, we generalise:
- `tl.trans(x)` from 2 dimensions to >= 2 dimensions, matching the
behaviour of numpy's `x.mT`;
- `tl.dot(A, B)` from 2 or 3 dimensions to >= 2 dimensions, supporting
multidimensional batches;
such that they batch over the first `n - 2` dimensions and operate on
the last 2 dimensions.

The generalised `tl.trans(x)` is useful for implementing the derivative
of `tl.dot`, as the same code can be written irrespective of the number
of batch dimensions:

    C = tl.dot(A, B)
    A_grad = tl.dot(C_grad, tl.trans(B))
    B_grad = tl.dot(tl.trans(A), C_grad)

Unit tests are included which test the new functionality of both
`tl.dot` and `tl.trans`.