[AMD] Rework lowering of ds_read_tr (b8/b16)
This uses a similar lowering to ldmatrix where we model the instruction
directly using the LL instead of modelling it as a transformation of the
requested output LL.
This codepath is used for B8/B16 and also FP4 packed along K types.

The path for FP4 packed along M/N is still using the legacy way since it
requires an input/output shape change that isn't as straightforward.