[AMD] Add `amdgpu.async_wait` to explicitly represent number of async transactions (#8575)
`ttg.async_wait` counts the number of outstanding `ttg.commit_groups`.
However, when lowering to LLVM on AMD we require the number of
outstanding async intrinsics/final assembly instructions. The conversion
is already done by `UpdateAsyncWaitCnt` which modifies the `num` of
`ttg.async_wait` in place.
This PR introduces a new op `amdgpu.async_wait` to make the change in
semantics explicit in the IR.

`UpdateAsyncWaitCount` is moved to `TTGIR->LLVM` primarily to also
include in for `Gluon` kernels and we should always call it since it
will only have an effect if there are `ttg.async_wait` ops present in
the kernel.

To avoid membar changes this also adds a `ttgpu.LocalBarrier` after each
`amdgpu.async_wait`. Membar will respect the newly added barrier and
behave the same as for `ttg.async_wait`.