Fix intermediate IR during expandLoops (#8589)
1. Ensure the types of the operands to subi are always the same.
2. Set the builder location before inserting constants, otherwise they
end up being inserted in a location that does not dominate the uses.
Flatten the control flow so it is obvious we always create an op.
3. Avoid inserting two yields in the else block if the `if` does not
produce a result (since the builder automatically inserts one in that
case). Simplify the code to just create both blocks ourselves.

Add an assert that the intermediate IR is valid, which currently fails
on existing lit tests.

I haven't constructed an example where these survive the pipeliner, but
it seems desirable to address these either way. It makes it less
confusing to inspect the IR during pipelining, and the invalid IR may
break things in unexpected ways in the future. (e.g. in the call to
applyPatternsGreedily in resolveMaskOps)