[CONSAN] Convert consan instrumentation opcodes to function calls (#8559)
This change aims reduces compilation time of kernels instrumented with
ConSan. It moves the implementation of the primitives from MLIR ops to
function calls. Only few of crucial ops are left as operations of
TritonInstrument dialect.
TritonInstrument::FunctionBuilder class is introduced as a class used
for creating and managing instrumentation functions.