[AMD][GLUON] Wait outstanding async commit groups instead of instructions (#8605)
Currently `async_wait` in Gluon on `CDNA4` requires the kernel writer to
pass the number of outstanding hardware instructions/llvm intrinsic to
`async_wait`. This count is very difficult to compute as it relies on
layouts, sizes, contiguity...

This PR changes the semantics of `async_wait` to represent the number of
outstanding commit groups. This follows the semantics used for nvidia in
Gluon. Therefore, Gluon kernels need to commit outstanding async
operations via `commit_group` and then wait on them via `wait_group`. I
also adapted the names so existing Gluon kernels using the old semantics
error out.

`UpdateAsyncWaitCount` is extended to compute the number of outstanding
hardware instructions based on the number of oustanding commits groups.
Previously, it only worked on `async_waits` carrying tokens of the
commit groups which are not available when compiling a Gluon kernel.
This is done by walking the IR backwards following *all* possible
control flow paths and finding the smallest number of emitted
instructions for N outstanding commit groups.