Fix handling of 3D DotOp with M=1  (#8561)
We currently always look at index 1 in the shape to determine if it is
an outer product. However, for 3D operands, we actually have to look at
index 2. This causes us to incorrectly crash when supplied a 3D operand
where the first dimension is 1.

Eliminate the isOuter check entirely, since my understanding is that:
1. it is purely defensive, since we will just crash when isOuter is
   true and mmaLayout is non-null.
2. it does not disqualify all the invalid K values, so it might be
   confusing.