[BACKEND] Take indices rather than reps as inputs to `{smem,tmem}Load` (#8623)
The previous code was doing a bad job at trying to guess the CTA-level
size of the tile being lowered. Here, we completely give up and instead
ask the caller to provide the starting coordinates of the subtensor they
want to lower rather than the reps.

In the passing, we also switch the `tmemLoad` logic to use LinearLayouts

Fixes https://github.com/triton-lang/triton/issues/8606