Fix resolving epilogue mask ops in nested loops (#8632)
If we have nested loops where both are pipelined, the contents of
peeledMaskOps will be invalidated when the outer loop is expanded,
because the original outer loop will be deleted. This causes problems
when we manipulate them in resolveMaskOps.

To address this, pull out the handling of MaskOps in the epilogue and do
it right after we expand each loop. While here, simplify the code so we
just replace the entire MaskOp with a PoisonOp, instead of replacing
each of the operations it contains.