[BACKEND] Initial support for 2CTA mode in Gluon (#8644)
To do this, we allow dispatching to 2CTA mode automatically in gluon by
looking at the layouts. In the near future we will tighten more the
op verifier so that the user gets a nice error message if they provide a
pair of layouts that is not compatible with regular mma or 2CTA

We also fix the layout for M=64 2CTA mode which was previously wrong.

In a follow-up, we will enable 2CTA mode to be launched with an
arbitrary number of CTAs.

Fixes https://github.com/triton-lang/triton/issues/8626