[GLUON] Implement `print_layout`
We do not want to expose layouts in gluon but we want to expose
encodings. That being said, we do want to offer the user a way to look
at the LinearLayout associated to a given encoding. This can be done by
using `LinearEncoding` and `SharedLinearEncoding` through
`to_linear_layout` for distributed and shared encodings, but there is
no way to print linear layouts for TMEM.

This feature addresses this point.

It's a bit of an RFC, I'm happy to expose this in a different way, if
you prefer to go all the way and expose LinearLayouts in gluon, but I
think that a targeted function like this one may be just fine for now.