[FRONTEND] Attach lineno information to AugAssign nodes (#8703)
visit_AugAssign works by creating fake AST nodes for the binary
operation and the assignment and visiting those. Prior to this commit,
the nodes lack the correct lineno and col_offset so any errors raised
here don't include the helpful '^' pointer to the offending line of
code; this makes debugging painful.