[CONSAN] Compilation time improvement (#8689)
Consan-generated functions are cached by the signature type. So far the
buffer/mbarrier memdesc was part of the signature, and since types of
these generally will be different for different buffers being accessed
in the kernel, we were ending up with duplicated functions where the
only difference was the type of the memdesc.
This PR moves the memdesc to i64 cast to before the function call, so
functions always take i64 param, reducing the duplication and the code
size. This saves some compilation time.

H100 (no warp spcialization):

|  | base | with consan | optimizing memdesc_to_i64 |
| --- | --- | --- | --- |
| make_ttir | 4,503 | 6,359 | 6,347 |
| make_ttgir | 25,823 | 32,835 | 32,785 |
| make_llir | 249,535 | 2,609,209 | 1,926,120 |
| make_ptx | 124,729 | 808,051 | 642,082 |
| make_cubin | 393,597 | 9,313,567 | 8,659,112 |

GB200 (WS):

|  | base | with consan | optimizing memdesc_to_i64 |
| --- | --- | --- | --- |
| make_ttir | 5,120 | 5,063 | 5,322 |
| make_ttgir | 181,326 | 181,617 | 186,404 |
| make_llir | 261,414 | 9,591,225 | 7,421,986 |
| make_ptx | 66,653 | 2,208,114 | 1,878,720 |
| make_cubin | 250,249 | 57,137,738 | 41,390,814 |