[NFC] Remove all the redundant CTALayout information from LIT tests (#8704)
It's not fully NFC, but as if it were. Before, we relied on the
CTALayout information to populate the CTALayout correctly for NVMMA.
Now, we serialise and deserialise a `rank` attribute to this effect.

This will still be necessary when we move to representing CTALayouts
with LinearLayouts, as this will give the number of output dimensions of
the associates LinearLayout (even if empty).

Weirdly enough, I found that on `ws-code-partition.mlir` there are some
`nvmma` layouts with `rank = 1`. I think those are wrong, but I didn't
want to keep touching things.