Preserve constexprs returned from functions
It is very easy to convert a constexpr into a tensor but difficult to go
in the reverse direction. The current behavior makes it very difficult
to support certain patterns without ceremony. While technically a
compatibility break, I expect this to be rarely hit because many
operations (like assignment) decay constexprs to tensors anyway.