[Frontend] Un-deprecate min/max on scalar tensors (#8734)
Looking at internal uses, a lot of min/max calls are in indexing
calculations with integer scalars. I think this is reasonable since it's
kind of hidden from the user that scalar variables are also tensors.

This also adds support for n-way min/max on tensors, since this was
added to the constexpr path in the previous pr.