Make WarpSpecializePartitionsOp implement RegionBranchInterface
Existing IR traversals for things like dataflow analysis cannot
automatically descend into the non-default partitions of a
WarpSpecializeOp because we do not encode the edges to these regions
through RegionBranchInterface.

Changing the warpspec operations to reflect these edges allows us to
clean up code in the dataflow analyses that walks the IR looking for
`WarpSpecializePartitionsOp`, and populates its block args.