Add enable_reflect_ftz flag to NVIDIA backend
libdevice's ftz behavior is controlled via nvvm-reflect (as opposed to
the ptxas ftz flag). As such, there is currertly no control over this
option. This means that e.g. it is hard to generate triton kernels that
match stock pytorch for activation functions. While it is possible to
monkey patch this out, providing an api seems like the better solution.

Previous MR:
- https://github.com/triton-lang/triton/pull/4626

See also previous discussion such as:
- https://github.com/llvm/llvm-project/issues/131749
- https://github.com/triton-lang/triton/issues/5735
- https://github.com/triton-lang/triton/issues/1535
- https://github.com/triton-lang/triton/pull/