[LAYOUTS] Make CTALayout an honest-to-goodness LinearLayout (#8770)
We represent CTALayout as a LinearLayout all across the compiler.
To do so, we introduce a `CTAEncoding` class that holds the relevant
`LinearLayout`
and exposes the legacy methods `getCTAsPerCGA()`, `getCTASplitNum()` and
`getCTAOrder`.

We split out `CTAEncoding` and `TritonGPU_Attr` into their own files to
be able to have
`CTAEncoding` fully defined before `AttrInterfaces.h.inc` is included.
Otherwise it was
creating a circular dependency. I was not completely sure how to do
this, codex fully did this. If anyone knows how to write this in a more
concise way please feel free to submit a follow-up.

For now, we still expose a creation method `fromSplitParams` that
creates a `CTAEncoding`
from these legacy params. The goal here would be to either fully remove
it, or keep it, but
remove the legacy methods. These two in combination are problematic, as
creating a `CTAEncoding`
using `fromSplitParams` from its legacy methods may not return the
original `CTAEncoding`.

We use the `CTAEncoding::getDefault` method in as many places as
possible whenever it was equivalent to
the legacy code. This flags these places as **incorrect** for any
multicta code. We should fix these.

We also avoid renaming `CTALayout` as `CGALayout` (a more correct name)
to avoid creating too big a diff here.
I'll submit a follow-up NFC doing so.