Remove verification of tmem allocation size
This check can fail, because `supportMMA` does not validate that the
result of a dot must be able to fit in tmem before using MMAv5. This
then results in an assertion failure.

Instead of hardcoding this constraint in the compiler, remove the check
and allow the allocator to run out of tmem later.