[Frontend] Support returning constexpr from functions
Closes #8749

Instead of eagerly promoting constexpr returns to tensor, this
now defers the decision until we have processed the entire function and
can look at all return values.

If all return paths return the same constexpr value, we return a
constexpr but if there are differing values we will promote to tensor
automatically.