[NFC] Simplify populating axisinfo map (#8800)
A default constructed `AxisInfo` passed as an operand to
`AxisInfo::join` will always result in `join` returning the other
operand. This means that we can call `join` unconditionally even when
there is no existing entry in the map.

This collapses the three separate map lookups (the check, the join, and
the population) to just a single one.