[AMD] Remove OptimizeLDSUsage pass (#8282)
This PR removes all optimizations related to LDS usage through layout
conversions.
The reason is that the current heuristics are outdated and no longer
align with the principles of linear layouts.
If future workloads require LDS consumption to be optimized, this can be
addressed by post-processing general swizzling algorithm where some
bases from the vector basis are shifted into the reps basis.