[WS] Nested loop support (#8687)
@3gx has done the bulk of work recently to get the partition
representation and aref passes ready for nested loops. So this PR, which
enables nested loops e2e, can focus on `partition-schedule` changes.
Most of the diffs are from there and the new tests.

Two main changes are made to `partition-schedule`:
* Initial partitioning and propagation / optimization of the partitions
are now done recursively according to the loop nest
* Hoisting of tmem_alloc to the top level. The existing hoisting pass,
when applied to nested loops, puts the hoisted alloc / store inside the
outer loop. The tmem store before the inner loop is problematic for
performance, because it introduces unnecessary sync between the default
and the MMA partitions (since the clearing of accum by zeros is done by
the default partition). The zero clear can be skipped since the inner
loop start with `useD = false`. To handle the case where the inner loop
does not execute, tmem_store must not be removed - rather, tmem alloc
and store need to be hoisted to the top level, outside of a WS-ed loop.

Hoisting of tmem_alloc across loop nests needs to be done carefully when
the inner loop depends on the outer loop tile ID. If it doesn't, all
instances of the inner loop execute the same number of times - either 0
or > 0. In this case, hoisting the zero clear to the top level is
correct. If it does, some output tiles execute MMA at least once while
none for others. In this case, we need to keep the zero clear before the
inner loop. But if the kernel has `tl.assume` op which asserts that the
inner loop executes at least once, like [how it is done in the MoE
kernel](https://github.com/triton-lang/triton/blob/main/python/triton_kernels/triton_kernels/matmul_ogs_details/_p_matmul_ogs.py#L260),
hoisting is still valid. We check that condition for simple cases of
`tl.assume` op over all outer loop iterators.

We will follow up with another PR to enable nested loop in the MoE
kernel.