[CONSAN] BufferRegion Analysis (#8837)
This change factors out the responsibility for determining what buffer
regions can be independently accessed in the program to a dedicated
analysis. Region is a pair (offset, length), and single shmem/tmem
allocation can be partitioned in many ways using indexing and slicing
operations. Before this change ConSan only did a very superficial IR
tracking to check if an allocation is used in a multi-buferred way.
Generating full list of the possible buffer regions is a pre-requisite
for supporting aliased shmem/tmem in ConSan. In the next step I will add
creating an aliasing matrix and proper checks to ConSan.
This change also introduces some smaller cleanups and refactoring of
FunctionBuilder, like moving the checks for existing auxData to the
`create` methods, and ditching error-prone `operator[]`.