[release/3.6] [Revert] [Coalesce] Fix the default order to be row major (#5707) #7143 (#8920)
Duplicates the work in
https://github.com/triton-lang/triton/commit/2d06bd2a03130d40a00b49c8db0ecfbb64989bab
for the 3.5 release to remove the change that makes the default ordering
row major. This is due to this PyTorch issue
https://github.com/pytorch/pytorch/issues/169492#issuecomment-3612342260,
which is a hang encountered in certain tests due to this change.

At this time it unclear if this is a bug in the Triton implementation or
just an incompatibility in implicit layout assumptions in the generated
Triton code for the reduction. Either way this will let us unblock the
issues in the release while we investigate the proper fix across the two
code bases.