[CUDA][PTXAS] Make TRITON_PTXAS_BLACKWELL_PATH work to overwrite ptxas-blackwell binary (#8945)
Mainly for THOR as by default ptxas-blackwell (cuda 12.9) would be used
and it would go "I don't understand sm110a".

Context: There is an environment variable TRITON_PTXAS_BLACKWELL_PATH
that was designed for users to overwrite this ptxas-blackwell, via the
lines noted in
https://github.com/triton-lang/triton/pull/8945#pullrequestreview-3562216960.
But: `export TRITON_PTXAS-BLACKWELL_PATH=/usr/local/cuda/bin/ptxas` is
NOT a valid bash script, but
`export TRITON_PTXAS_BLACKWELL_PATH` is. So this PR maps
`ptxas-blackwell` binary to look for `TRITON_PTXAS_BLACKWELL_PATH`. IF
set, the ptxas-blackwell would use the provided "path" (nit: it is call
a "path" but we should supply the binary, like
"TRITON_PTXAS_BLACKWELL_PATH_AND_BINARY".