[CONSAN] Adding support for aliasing (#8939)
This change adds support for buffer aliasing in ConSan. This is achieved
by using BufferRegion analysis to create AliasingMatrix that marks which
buffers may overlap with each other. This information is then used
during the visibility checks: instead of verifying that the buffer being
accessed is visible, we load a row from the aliasing matrix, describing
all the buffers that may alias buffer being access (including the buffer
in question). This set of buffers is then checked for having visibility
into potential outstanding reads/writes.

There is still follow-up change coming: moving BufferRegion analysis to
TritonInstrument. It was not done in this PR to make review easier.