[Backend] Consider memdesc_slice in the Membar analysis (#8755)
Instead of always using the full allocation for each operation we can
take the subslices into account, if any are present.
Using this mechanism we can tighten memory accesses and it can result
in fewer local_barrier being inserted.

The intersection of the accessed memory is now based on
- allocation interval
- access type (shape, layout)
- subslice offsets

The current intersection check works by comparing intervals, if the
intervals intersect
then it will do a more in depth analysis the subslice offsets (if
present) to determine
if the two subslices intersect.

If access layouts don't match, we assume intersection. This layout
constraint can be
relaxed in the future if needed.