[Gluon][Dialect] Tighten verifiers, add more helpful error messages (#8981)
`DotOpMMASmemLoader::build` is now fallible if it cannot find an SMEM
atom to implement the layout. Since the logic is fairly complex, it
perhaps doesn't make sense as a verifier. This is piped through the
lowering of `WarpGroupDotOp`, `TMEMCopyOp`, and the MMAv5 ops, which are
not fallible too.

* Add verifiers to `tma` functions in Gluon, especially `async_gather`
and `async_scatter` which will happily trigger illegal instruction if
runtime invariants are not satisfied
* Verify `NVMMASharedEncodingAttr` is valid and fix all the unit tests
* Misc other cleanups

Some of these should eventually get moved up to Python for better UX but
it's sufficient for the error message to be more actionable.