[NFC] Improve robustness of ext slice rematerialization
When trying to extend the instructions to rematerialise to include the
backward slice of the ext, we currently perform the following steps:
1. Attempt to obtain the backward slice of the ext independently
2. Validate that the slice can be safely merged with the current slice
3. Merge the slices

Instead, we can attempt to directly extend the existing slice and just
roll back if that fails. This is potentially slightly less efficient,
since we copy the slice and layout, but it means that the validation
happens entirely within `getConvertBackwardSlice`.

This organisation would have sidestepped the bug fixed in #7058, and the
bug that caused #8292 to be reverted.

While investigating this, I have found that the code for hoisting into
conditionals may be prone to the same bug, since it also attempts to
extend an existing slice. But I will keep that for a separate PR since
it is more complex and I need to write a test case.