[Reapply][BACKEND] Add missing waits in WGMMA rhs in register pipelining (#8997)
This reverts #8970 with an additional fix to the non-persistent matmul
kernel that fixes wgmma pipelining for bf16xmxfp4.

The performance regression was caused by the use of `acc.trans()` inside
the loop which is seen as a use of the accumulator and triggers a wait 0
to be generated. The fix is to use the same pattern as `_p_matmul` and
hoist the transpose out of the loop.