Move gluon changes from #8981 into C++ verifiers
My attention was drawn to these changes because the verifier is too strict,
in particular
```python
assert tensor_desc.layout == smem.layout
```
breaks dimension-reducing loads, which have different ranks between the
descriptor and destination memory.

The C++ verifier handles this correctly, and is better tested by virtue
of being used in the normal triton compilation flow. So, it's my opinion
that we should use the C++ verifiers as the main source of truth.