[CONSAN] Fix for false positive deadlock detection (#9046)
ConSan was ignoring barrier_expect, instead treating every
`tma.async_copy_global_to_shared` as always arriving on a barrier.
Because of this, if two or more `tma.async_copy_global_to_shared` were
sharing a barrier, ConSan would model this as barrier overarrival and
report as a deadlock if such barrier is waited on.
Ultimate fix for this would be to add full support for counting
transferred bytes and flipping barrier phase when it is being reached.
This PR proposes simpler solution, which is treating `mbarrier.expect`
as an arrive. This is more correct than behavior before the fix, but
still allows for deadlocks caused by wrong transfer size in `expect`.