[AMD] Refactor direct-to-lds warp coalescing check for GFX9 (#9048)
Refactors `canCoalesceWriteIntoSharedMemory` to just check if we can
divide the `srcToShared` layout by a identity layout for one instruction
(based on contig). This tells us if warps write coalesced to LDS which
is a hardware requirement on `GFX9`. This also makes
`doesSwizzleInsideWarp` redundant.

The lowering now also correctly rejects subslices which break the warp
coalesced writes (see added lit tests). Note that we pass the alloc
shape instead of the `memdesc` because as a follow up we will use this
function in the pipeliner where we do not have the complete memdesc when
selecting layouts.

---------

Co-authored-by: Lei Zhang <antiagainst@gmail.com>