[BACKEND] Fix uses of CGAEncodingAttr::getDefault (#9040)
`getDefault` as implemented, just made sense for 1CTA mode. In this PR
we split it into two helper functions `get1CTALayout` and `get1DLayout`.

We also add a verifier that checks that MemDescTypes have the right
amount
of bases in their cga_layout/block attribute.

We also fix the CoalesceUtils that didn't use to support multicta, and
now they do. I will exercise this in a follow-up PR.