[WS] Correctly assign stage / cluster annotation to aref ops for a block-arg producer (#8883)
When the "produced value" is a block argument, such as the "row-max
tensor" in attention, we are not assigning stage / cluster annotations
to aref put ops and `local_store`. This is a bug, but it didn't manifest
for fused loop / non-persistent kernel because the second
`schedule-loop` pass that runs after WS annotates them. For nested-loop
persistent attention, such "fixup" didn't work for some reason, so we
end up getting (benign) errors from the pipeliner like `<unknown>:0:
error: 'arith.addi' op not assigned a pipeline stage`.

`RewritePartitionDependices` was correctly handling this case, thanks to
this logic
https://github.com/triton-lang/triton/blob/de8e71503fea971dfb65308147798657e18f8568/lib/Dialect/TritonGPU/Transforms/WarpSpecialization/RewritePartitionDependencies.cpp#L231-L240.
Porting that to `insert-aref` fixes this issue.